+++ src/ncsh_autocompletions.h
@@ -29,23 +29,23 @@ struct ncsh_Autocompletion_Node* ncsh_autocompletions_malloc();
 
 void ncsh_autocompletions_free(struct ncsh_Autocompletion_Node* tree);
 
-void ncsh_autocompletions_add(char* string, uint_fast32_t length, struct ncsh_Autocompletion_Node* tree);
+void ncsh_autocompletions_add(char* string, size_t length, struct ncsh_Autocompletion_Node* tree);
 void ncsh_autocompletions_add_multiple(struct eskilib_String* strings, uint_fast32_t count, struct ncsh_Autocompletion_Node* tree);
 
-struct ncsh_Autocompletion_Node* ncsh_autocompletions_search(char* string, uint_fast32_t length, struct ncsh_Autocompletion_Node* tree);
+struct ncsh_Autocompletion_Node* ncsh_autocompletions_search(char* string, size_t length, struct ncsh_Autocompletion_Node* tree);
 struct ncsh_Autocompletion_Node* ncsh_autocompletions_search_string(struct eskilib_String string, struct ncsh_Autocompletion_Node* tree);
 
 // gets all matches based on traversing the tree.
 // populates matches into variable matches and returns 0 if no matches, number of matches length if any matches.
 uint_fast8_t ncsh_autocompletions_get(char* search,
-				       uint_fast32_t search_length,
+				       size_t search_length,
 				       struct ncsh_Autocompletion* matches,
 				       struct ncsh_Autocompletion_Node* tree);
 
 // gets highest weighted match based on traversing the tree.
 // populates match into variable match and returns 0 if not matches, 1 if any matches.
 uint_fast8_t ncsh_autocompletions_first(char* search,
-				       uint_fast32_t search_length,
+				       size_t search_length,
 				       char* match,
 				       struct ncsh_Autocompletion_Node* tree);
